Moved upload_file_to_bucket_fs.py to correct submodule
